backend/requirements.txt
django>=5.2
djangorestframework>=3.15
django-cors-headers>=4.3
dj-database-url>=2.1
psycopg2-binary>=2.9
python-dotenv>=1.0
PyJWT>=2.8
cryptography>=42.0
pytz>=2024.1
requests>=2.31
stripe>=8.0
gunicorn>=21.2
whitenoise>=6.6
